third_party/blink/web_tests/fast/js/mozilla/strict/eval-variable-environment-expected.txt

PASS 'use strict'; eval('var x = 2; typeof x'); is "number"
PASS eval('var x = 2; typeof x'); is "number"
PASS true === true
PASS 'use strict'; eval('"use strict"; var x = 2; typeof x'); is "number"
PASS eval('"use strict"; var x = 2; typeof x'); is "number"
PASS true === true
PASS 'use strict'; eval('var x = 2;'); typeof x is "undefined"
PASS eval('var x = 2;'); typeof x is "number"
PASS true === true
PASS 'use strict'; eval('"use strict"; var x = 2;'); typeof x is "undefined"
PASS eval('"use strict"; var x = 2;'); typeof x is "undefined"
PASS true === true
PASS 'use strict'; eval('"use strict"; var x = 2; typeof x'); typeof x is "undefined"
PASS eval('"use strict"; var x = 2; typeof x'); typeof x is "undefined"
PASS true === true
PASS 'use strict'; function test() {   eval('var x = 2;');   return typeof x; } test(); is "undefined"
PASS function test() {   eval('var x = 2;');   return typeof x; } test(); is "number"
PASS true === true
PASS 'use strict'; function test() {   'use strict';   eval('var x = 2;');   return typeof x; } test(); is "undefined"
PASS function test() {   'use strict';   eval('var x = 2;');   return typeof x; } test(); is "undefined"
PASS true === true
PASS 'use strict'; function test() {   eval('"use strict"; var x = 2;');   return typeof x; } test(); is "undefined"
PASS function test() {   eval('"use strict"; var x = 2;');   return typeof x; } test(); is "undefined"
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
